pyproject.toml
[project]
name = "realfill"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.6.0",
    "bitsandbytes>=0.45.5",
    "diffusers>=0.33.1",
    "ftfy>=6.3.1",
    "huggingface-hub>=0.31.1",
    "jinja2>=3.1.6",
    "kornia>=0.8.1",
    "peft>=0.15.2",
    "tensorboard>=2.19.0",
    "torch>=2.7.0",
    "torchvision>=0.22.0",
    "transformers>=4.51.3",
    "xformers>=0.0.30",
]

[project.optional-dependencies]
benchmark = [
    "dreamsim>=0.2.1",
    "kornia>=0.8.1",
    "lpips>=0.1.4",
    "matplotlib>=3.10.1",
    "numpy>=2.2.5",
    "opencv-python>=4.11.0.86",
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "rich>=14.0.0",
    "timm>=1.0.15",
    "torch>=2.7.0",
    "torchvision>=0.22.0",
    "tqdm>=4.67.1",
    "transformers>=4.51.3",
]

[tool.black]
line-length = 100
skip-magic-trailing-comma = true

[tool.isort]
profile = "black"

[dependency-groups]
dev = [
    "black[jupyter]>=25.1.0",
    "pylint>=3.3.7",
]
